requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
boto3>=1.34.0
pyyaml>=6.0
pyyaml
